Move PAINT's method modules inside engines/paint
