+++ src/modules/highlight/engines/paint/method.mts
