+++ src/modules/highlight/engines/paint/methods/paint.mts
@@ -1,4 +1,4 @@
-import type { Box, AbstractMethod } from "src/modules/highlight/method.mjs"
+import type { Box, AbstractMethod } from "src/modules/highlight/engines/paint/method.mjs"
 import type { AbstractHighlightability } from "src/modules/highlight/highlightability.mjs";
 const { EleID, EleClass } = await import("src/modules/common.mjs");
 
@@ -90,4 +90,4 @@ class CSSPaintHighlightability implements AbstractHighlightability {
 	}
 }
 
-export { PaintMethod }
+export { TermSelectorStyles, PaintMethod }
